Create the directory for a database immediatly
It seems redundant to create the directory in the loop; furthermore, I
think it's probably better to have the empty directory if the database
is empty, vs not creating the directory if the database is empty. Said
another way, it seems an empty database should correspond to an empty
directory.