Refactor block class constructor
Previously, the Block class constructor would iterate through all of the
top-level keys of the Notion block objects, dynamically adding
them all as attributes the the Block instance. Then, it would loop
through all of the block-specific keys and also add them as attributes.
Although this was working, in some cases the block-specific keys
would over-ride the top-level keys (e.g., the "type" key in particular
was being shadowed in a few cases).

Dynamically adding the keys makes it more difficult to read and debug the code,
since you don't know what keys are going to be set until run-time. This
makes writing plugins more difficult. Also, errors due to changes in the
notion API (or even issues with poorly constructed mock test data) don't
appear until the expected keys are used in the `to_pandoc` calls, rather
than during the constructors.

This commit explicilty sets all of the top-level keys. It also moves the
"type" key into the "notion_type" attribute, which avoids shadowing
issues if the block-specific object has a "type" key. The "id" key was
also moved to "notion_id".

After making this change, the unit tests were failing since now some of
the top-level keys that previously weren't being set in the unit tests
were now missing. To ensure that these keys are always present, the
tests were updated to use a set of mock objects that ensure the keys
always exist.

Furthermore, the tests related to the rich_text objects were split out
into their own file.

Duplication in the test was also removed through the addition of the
`process_block` and `process_parent_block` functions.

There is a second refactor which I believe should occur, which is to
move the block-specific attributes into each block's constructors. This
will make the block-specific attribute construction more explicit as
well, which is good for all of the reasons stated above. A bit of work
in this direction was made, but more work is needed.