Handle property values using classes
Similar to how we handle all other database objects, we now handle
"PropertyValues" using classes too. One big benefit of this approach is
that the markdown rendering of rich text is all handled in one place.
Previously there was a half-baked rich text processing system in the
"property_values.py" file.

Add a bunch of unit tests for various property values, as well as one
big end-to-end test that uses most (if not all) propertly values.

There are still a few small edge-cases that need to be taken care of,
such as formulas that produce arrays of dates, but it's pretty close to
complete.

Start adding a mechanism to save a set of related databases as YAML
files; this doesn't work due to an apparent bug in Notion. In
particular, the current code identifies related databases using the
`database.schema` property, instead of looking at the property values in
pages in that database. Using the database schema would be preferred
since it will work even if you have an empty database. It will also be
faster. However, it appears that Notion's API is not returning the
"property object" for relations that point to other databases. (In
particular, the "Employer" relation.) It _does_ include the
self-relation ("Mom"). I will look into this issue further and will
upgrade the API version we're using, but then I'll report this to Notion
to see if they can fix it.