Fix a bug in the audit system
The `Page.children` property was not recursing into child blocks and as
a result any child database or page that was not a top-level block would
be picked up by the audit system.

This bug would affect other things too, although it was noticed due to
an audit producing false positives.