Simple updates (#116)
* Remove warning for database pages

It used to be that these were always un-intentional, but now we often
intentionally put databases inline so that content can be reviewed in
notion. We usually have a jinja render block right below it. Eventually,
when notion support linked database views, we will probalby implement
this.

* Add more debug information to jinja error printouts

Perhaps it would be good to add some values too ...

* Make the default join key "notion_id"

This is what we will usually use moving forward

* Fix flake8 error

* Fix broken test